Add test setup for CreateThumbnail Function
