+++ spec/user_form/confirm_upload_spec.rb
@@ -1,15 +1,6 @@
 require "spec_helper"
 require_relative "../../user_form/confirm_upload"
-
-class FakeLambdaContext
-  attr_accessor :function_name, :function_version, :invoked_function_arn,
-    :memory_limit_in_mb, :aws_request_id, :log_group_name, :log_stream_name,
-    :deadline_ms, :identity, :client_context
-
-  def get_remaining_time_in_millis
-    3000
-  end
-end
+require_relative "../fakes/lambda_context"
 
 RSpec.describe UserForm::ConfirmUpload do
   let(:confirm_upload_event) {
