+++ spec/user_form/show_form_spec.rb
@@ -1,16 +1,7 @@
 require "spec_helper"
 require "json"
 require_relative "../../user_form/show_form"
-
-class FakeLambdaContext
-  attr_accessor :function_name, :function_version, :invoked_function_arn,
-    :memory_limit_in_mb, :aws_request_id, :log_group_name, :log_stream_name,
-    :deadline_ms, :identity, :client_context
-
-  def get_remaining_time_in_millis
-    3000
-  end
-end
+require_relative "../fakes/lambda_context"
 
 RSpec.describe UserForm::ShowForm do
   let(:show_form_event) {
